dashboard/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0